[wasm]Add broadcast to all binary ops.
