+++ tfjs-core/src/ops/binary_ops_test.ts
@@ -1273,23 +1273,3 @@ describeWithFlags('div', ALL_ENVS, () => {
     expectArraysClose(await c.data(), [0, 0, 0, 0]);
   });
 });
-
-describeWithFlags('floorDiv', ALL_ENVS, () => {
-  it('floorDiv', async () => {
-    const a = tf.tensor1d([10, 20, -20, -40], 'int32');
-    const b = tf.tensor1d([10, 12, 8, 5], 'int32');
-    const result = tf.floorDiv(a, b);
-
-    expect(result.shape).toEqual(a.shape);
-    expectArraysClose(await result.data(), [1, 1, -3, -8]);
-  });
-
-  it('floorDiv vec4', async () => {
-    const a = tf.tensor1d([10, 20, -20, -40], 'int32');
-    const b = tf.tensor1d([10, 12, 8, 5], 'int32');
-    const result = tf.floorDiv(a, b);
-
-    expect(result.shape).toEqual(a.shape);
-    expectArraysClose(await result.data(), [1, 1, -3, -8]);
-  });
-});
